Homework8

+++ app/src/main/java/ru/dvn/moviesearch/model/movie/detail/remote/DetailsApi.kt
@@ -4,13 +4,11 @@ import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.Path
-import ru.dvn.moviesearch.model.movie.detail.MovieDetailDto
-import ru.dvn.moviesearch.model.movie.list.FilmDTO
 
 interface DetailsApi {
     @GET("api/v2.2/films/{id}")
     fun getFilmDetail(
         @Header("X-API-KEy") key: String,
-        @Path("id") id: Int
+        @Path("id") id: Long
     ): Call<MovieDetailDto>
 }
